Separate Azure CloudManager functionality into mixin
Separate Azure CloudManager functionality into mixin so it's
usable in NetworkManager
